feat: areas store module + includes normalizer
