+++ package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "file-saver": "^1.3.3",
     "history": "^4.6.1",
+    "json-api-normalizer": "^0.4.0",
     "prop-types": "^15.5.10",
     "query-string": "^4.3.4",
     "react": "^15.5.4",
