+++ src/modules/index.js
@@ -2,3 +2,4 @@
 export { default as app } from './app';
 export { default as user } from './user';
 export { default as data } from './data';
+export { default as areas } from './areas';
